+++ CHANGELOG.md
@@ -10,6 +10,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Fixed
 - Fixed the_block_content function when using a string as post_id.
 
+### Changed
+-Changed instantpage script to load from plugin instead of outside source.
+
 ## [2.11.0] & [2.11.1] 2021-09-23
 ### Added
 - Remove plugins page from multisite admin menu